requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gradio==4.19.2
websockets==12.0
sqlalchemy==2.0.25
qdrant-client==1.7.0
httpx==0.26.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
rsocket-py==0.5.1
aiohttp==3.9.1
reactivestreams==1.0.4
numpy==1.26.3
soundfile==0.12.1
torch==2.1.2
tqdm==4.66.1
